requirements.txt
google-generativeai
pandas
numpy
scikit-learn
datasets
tqdm
python-dotenv
pyyaml
openai